avoid referencing entity classes while patching
prevents an issue that would prevent sponge from being able to process
it’s patches. I can imagine this would also effect other coremods
trying to patch EntityPlayer
