+++ src/main/java/com/mordenkainen/bedpatch/BedPatchASM.java
@@ -1,8 +1,5 @@
 package com.mordenkainen.bedpatch;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.objectweb.asm.ClassWriter;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.tree.AbstractInsnNode;
@@ -16,12 +13,7 @@ import org.objectweb.asm.tree.VarInsnNode;
 import com.mordenkainen.bedpatch.asmhelper.ASMHelper;
 import com.mordenkainen.bedpatch.asmhelper.ObfHelper;
 
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.launchwrapper.IClassTransformer;
-import net.minecraft.util.ClassInheritanceMultiMap;
-import net.minecraft.world.World;
-
 
 public class BedPatchASM implements IClassTransformer {
 
@@ -36,7 +28,7 @@ public class BedPatchASM implements IClassTransformer {
             insnList.add(new FieldInsnNode(Opcodes.GETFIELD, "net/minecraft/world/chunk/Chunk", ObfHelper.isObfuscated() ? "field_76637_e" : "world", "Lnet/minecraft/world/World;"));
             insnList.add(new VarInsnNode(Opcodes.ALOAD, 0));
             insnList.add(new FieldInsnNode(Opcodes.GETFIELD, "net/minecraft/world/chunk/Chunk", ObfHelper.isObfuscated() ? "field_76645_j" : "entityLists", "[Lnet/minecraft/util/ClassInheritanceMultiMap;"));
-            insnList.add(new MethodInsnNode(Opcodes.INVOKESTATIC, "com/mordenkainen/bedpatch/BedPatchASM", "onChunkUnload", "(Lnet/minecraft/world/World;[Lnet/minecraft/util/ClassInheritanceMultiMap;)V", false));
+            insnList.add(new MethodInsnNode(Opcodes.INVOKESTATIC, "com/mordenkainen/bedpatch/BedPatchFunc", "onChunkUnload", "(Lnet/minecraft/world/World;[Lnet/minecraft/util/ClassInheritanceMultiMap;)V", false));
             AbstractInsnNode insert = ASMHelper.findFirstInstruction(method);
             method.instructions.insertBefore(insert, insnList);
             
@@ -44,18 +36,6 @@ public class BedPatchASM implements IClassTransformer {
         }
         
         return basicClass;
-    }
-    
-    public static void onChunkUnload(World world, ClassInheritanceMultiMap<Entity>[] entityLists) {
-        List<EntityPlayer> players = new ArrayList<EntityPlayer>();
-        for (ClassInheritanceMultiMap<Entity> classinheritancemultimap : entityLists) {
-            for(EntityPlayer player : classinheritancemultimap.getByClass(EntityPlayer.class)) {
-                players.add(player);
-            }
-        }
-        for (EntityPlayer player : players) {
-            world.updateEntityWithOptionalForce(player, false);
-        }
     }
     
 }
